+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Deprecated `Config.enviroment` (typo) in favor of `Config.environment`.
+- Clarified what `Config.environment` affects.
+
 ### Removed
 
 - CDR and PDF support in Sanitize.